backend/pyproject.toml
[tool.poetry]
name = "sharegratitude"
version = "0.1.0"
description = ""
authors = ["gwonin-kenji <hangwonin@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.74.1"
uvicorn = "^0.17.5"
psycopg2 = "^2.9.3"
SQLAlchemy = "^1.4.31"
email-validator = "^1.1.3"
passlib = "^1.7.4"
bcrypt = "^3.2.0"
python-dotenv = "^0.19.2"
requests = "^2.27.1"
python-jose = "^3.3.0"

[tool.poetry.dev-dependencies]
icecream = "^2.1.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
